requirements.txt
rdflib
pyvis
python-dotenv
google-genai
transformers
torch
sentencepiece
json-repair>=0.59.4
networkx>=3.6.1
langchain-core>=1.2.31
langchain-google-vertexai>=3.2.2
langchain-experimental>=0.4.1
langchain-community>=0.4.1
rapidfuzz
